Move everything to a single file so it's easy to copy+paste. Add a GitHub action to show things work
